% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.mean.R
\name{ci.mean}
\alias{ci.mean}
\title{Compute mean values with confidence intervals}
\usage{
ci.mean(x, ...)
}
\arguments{
\item{x}{object passed to methods}

\item{...}{passed to methods}
}
\value{
a list with mean values and confidence limits
}
\description{
Compute mean values with confidence intervals
}
\details{
Normal approximation
}
