% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org.R
\name{org}
\alias{org}
\title{Wrapper function for publish with output format org}
\usage{
org(x, ...)
}
\arguments{
\item{x}{object to format as org}

\item{...}{passed to publish}
}
\value{
See publish
}
\description{
Wrapper for \code{publish(...,org=TRUE)}
}
\author{
Thomas Alexander Gerds
}
