% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.subgroupAnalysis.R
\name{plot.subgroupAnalysis}
\alias{plot.subgroupAnalysis}
\title{plot.subgroupAnalysis}
\usage{
function(fit,subgroupVar="subgroupVar",SubgroupVal="SubgroupVal",
Treatment="Treatment",...)
}
\arguments{
\item{fit}{- a subgroupAnalysis object}

\item{subgroupVar}{- name for column with variables forming subgroups}

\item{SubgroupVal}{- name of column for levels of subgroupVar}

\item{Treatment}{- name of variable that defines subgroup levels}
}
\value{
A plot
}
\description{
This function operates on a "subgroupAnalysis" object to produce a formatted
table and a forest plot
}
\details{
This function produces a formatted table of a subgroupAnalysis object and
adds a forest plot.  If further details needs attention before plotting is
is advisable use adjust the table produced by thesummary function and then
plotting with the plotConfidence function
}
\examples{
#load libraries
library(Publish)
library(survival)
data(traceR) #get dataframe traceR
traceR$wmi2 <- ifelse(traceR$wallMotionIndex<0.9,"bad","good")
traceR$abd2 <- ifelse(traceR$abdominalCircumference<95,"slim","fat")
# Selected subgroups - univariable analysis
sub <- subgroupAnalysis(traceR,outcome="dead",timevar="observationTime",
  treatment="treatment", subgroup=c("smoking","sex","wmi2","abd2")) 
plot(sub)  
}
\seealso{
subgroupAnalysis, plotConfidence
}
\author{
Christian Torp-Pedersen
}
