% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotConfidence.R
\name{plotConfidence}
\alias{plotConfidence}
\title{Plot confidence intervals}
\usage{
plotConfidence(x, lower, upper, labels, title.labels, values, title.values,
  order, factor.reference.pos, factor.reference.label = "Reference",
  factor.reference.pch = 8, xratio = 0.618, y.offset = 0,
  y.title.offset = 1.3, pch = 16, cex = 1, lwd = 1, col = 1,
  digits = 2, format = "(-.-)", value.format, xlim, xlab, add = FALSE,
  axes = TRUE, ...)
}
\arguments{
\item{x}{Either a vector containing the point estimates or A list
whose first element contains the point estimates.  Further list
elements can contain the confidence intervals and labels. In this
case the list needs to have names 'lower' and 'upper' to indicate
the values of the lower and the upper limits of the confidence
intervals, respectively, and 'labels' to indicate a vector or
matrix or list with labels.}

\item{lower}{Lower confidence limits. Used if object \code{x} is a
vector and if \code{x} is a list \code{lower} overwrites element
\code{x$lower}.}

\item{upper}{Upper confidence limits. Used if object \code{x} is a
vector and if \code{x} is a list \code{upper} overwrites element
\code{x$upper}.}

\item{labels}{Vector or matrix or list with \code{labels}. Used if
object \code{x} is a vector and if \code{x} is a list it
overwrites element \code{x$labels}. If \code{labels=FALSE} do not
draw labels.}

\item{title.labels}{Title of the \code{labels}. If \code{labels}
is a matrix or list \code{title.labels} should be a vector with as
many elements as labels has columns or elements.}

\item{values}{Either logical or vector, matrix or list with
values. If \code{values=TRUE} values are constructed according to
\code{format} from \code{lower} and \code{upper} overwrites
constructed values. If \code{values=FALSE} do not draw values.}

\item{title.values}{Title of the \code{values}. If \code{values}
is a matrix or list \code{title.labels} should be a vector with as
many elements as values has columns or elements.}

\item{factor.reference.pos}{Position at which factors attain
reference values.}

\item{factor.reference.label}{Label to use at
\code{factor.reference.pos} instead of values.}

\item{factor.reference.pch}{Plotting symbol to use at \code{factor.reference.pos}}

\item{xratio}{}

\item{y.offset}{}

\item{y.title.offset}{}

\item{pch}{}

\item{cex}{}

\item{lwd}{}

\item{col}{}

\item{digits}{}

\item{format}{}

\item{value.format}{}

\item{xlim}{}

\item{xlab}{}

\item{add}{}

\item{axes}{}

\item{...}{}
}
\value{
List of coordinates
}
\description{
Function to plot confidence intervals
}
\details{
Function to plot means and other point estimates with confidence intervals
}
\examples{
library(Publish)
data(CiTable)
plotConfidence(x=CiTable[,6:8],
               labels=CiTable[,1:5])



plotConfidence(x=CiTable[,6:8],
               labels=CiTable[,1:5],
               xlab="Hazard ratio",
               title.values=expression(bold(HR (CI[95]))),
               factor.reference.pos=c(1,10,19),
               cex=1.3,
               arrows.length=0.05,
               axis1.at=c(0.75,1,1.25,1.5,2))

## there are three blocks: labels, confidence intervals, values
## the order of appearance can be changed
plotConfidence(x=CiTable[,6:8],
               labels=CiTable[,1:5],
               xlab="Hazard ratio",
               factor.reference.pos=c(1,10,19),
               cex=1.3,
               order=c(1,3,2),
               arrows.length=0.05,
               axis1.at=c(0.75,1,1.25,1.5))

## changing the style of the graphical confidence intervals
plotConfidence(x=CiTable[,6:8],
               labels=CiTable[,1:5],
               xlab="Hazard ratio",
               factor.reference.pos=c(1,10,19),
               points.pch=15,
               points.cex=2,
               arrows.col="darkblue",
               cex=1.3,
               order=c(1,3,2),
               axis1.at=c(0.75,1,1.25,1.5))


## the values can have multiple columns as well
## for this we create the confidence intervals
## before calling the function and then cbind them
## to the pvalues
CI95 <- formatCI(lower=CiTable[,7],upper=CiTable[,8],format="(u-l)")
pval <- format.pval(CiTable[,9],digits=3,eps=10^{-3})
pval[pval=="NA"] <- ""
plotConfidence(x=CiTable[,6:8],
               labels=CiTable[,1:5],
               values=list("CI-95"=CI95,"P-value"=pval),
               xratio=c(0.4,0.4))
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}

