% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.subgroupAnalysis.R
\name{print.subgroupAnalysis}
\alias{print.subgroupAnalysis}
\title{Printing univariate tables}
\usage{
\method{print}{subgroupAnalysis}(x, ...)
}
\arguments{
\item{x}{- An object obtained with \code{subgroupAnalysis}}

\item{...}{Passed to summary.subgroupAnalysis}
}
\value{
The result of \code{summary.subgroupAnalysis(x)}
}
\description{
Print function for subgroupAnalysis
}
\details{
This function is simply calling \code{summary.subgroupAnalysis}
}
\seealso{
\code{subgroupAnalysis}
}
\author{
Christian Torp-Pedersen (ctp@heart.dk)
}
