% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.table2x2.R
\name{print.table2x2}
\alias{print.table2x2}
\title{print results of 2x2 contingency table analysis}
\usage{
\method{print}{table2x2}(x, digits = 1, ...)
}
\arguments{
\item{x}{object obtained with table2x2}

\item{digits}{rounding digits}

\item{...}{not used}
}
\value{
invisible x
}
\description{
print results of 2x2 contingency table analysis
}
\examples{
table2x2(table("marker"=rbinom(100,1,0.4),"response"=rbinom(100,1,0.1)))
table2x2(matrix(c(71,18,38,8),ncol=2),stats="table")
table2x2(matrix(c(71,18,38,8),ncol=2),stats=c("rr","fisher"))
}
\seealso{
table2x2
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
