% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.univariateTable.R
\name{print.univariateTable}
\alias{print.univariateTable}
\title{Printing univariate tables}
\usage{
\method{print}{univariateTable}(x, ...)
}
\arguments{
\item{x}{An object obtained with \code{univariateTable}}

\item{...}{Passed to summary.univariateTable}
}
\value{
The result of \code{summary.univariateTable(x)}
}
\description{
Print function for univariate tables
}
\details{
This function is simply calling \code{summary.univariateTable}
}
\seealso{
\code{univariateTable}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
