% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.htest.R
\name{publish.htest}
\alias{publish.htest}
\title{Pretty printing of test results.}
\usage{
\method{publish}{htest}(object, title, ...)
}
\arguments{
\item{object}{Result of \code{t.test} or \code{wilcox.test}}

\item{title}{Decoration also used to name output}

\item{...}{Used to transport arguments \code{ci.arg} and \code{pvalue.arg} to subroutines \code{format.pval} and \code{formatCI}. See also \code{prodlim::SmartControl}.}
}
\description{
Pretty printing of test results.
}
\examples{
data(Diabetes)
publish(t.test(bp.2s~gender,data=Diabetes))
publish(wilcox.test(bp.2s~gender,data=Diabetes))
publish(with(Diabetes,t.test(bp.2s,bp.1s,paired=TRUE)))
publish(with(Diabetes,wilcox.test(bp.2s,bp.1s,paired=TRUE)))

}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
