% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publishing tables and figures}
\usage{
publish(object, ...)
}
\arguments{
\item{object}{object to be published}

\item{...}{Passed to method.}
}
\value{
Tables and figures
}
\description{
Publish provides summary functions for data
and results of statistical analysis in ready-for-publication
design
}
\details{
Some warnings are currently suppressed.
}
\author{
Thomas Gerds
}

