% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.summary.aov.R
\name{publish.summary.aov}
\alias{publish.summary.aov}
\title{Format summary table of aov results}
\usage{
\method{publish}{summary.aov}(object, print = TRUE,
  handler = "sprintf", digits = c(2, 4), nsmall = digits, ...)
}
\arguments{
\item{object}{glm object}

\item{print}{Logical. Decide about whether or not to print the results.}

\item{handler}{see \code{pubformat}}

\item{digits}{see \code{pubformat}}

\item{nsmall}{see \code{pubformat}}

\item{...}{used to transport further arguments}
}
\description{
Format summary table of aov results
}
\examples{
 
data(Diabetes)
f <- glm(bp.1s~age+chol+gender+location,data=Diabetes)
publish(summary(aov(f)),digits=c(1,2))

}
