% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.survdiff.R
\name{publish.survdiff}
\alias{publish.survdiff}
\title{Alternative summary of survdiff results}
\usage{
\method{publish}{survdiff}(object, digits = c(2, 4), print = TRUE, ...)
}
\arguments{
\item{object}{Object obtained with \code{survival::survdiff}.}

\item{digits}{Vector with digits for rounding numbers: the second for pvalues, the first for all other numbers.}

\item{print}{If \code{FALSE} do not print results.}

\item{...}{Not (yet) used.}
}
\description{
Alternative summary of survdiff results
}
\examples{
library(survival)
data(pbc)
sd <- survdiff(Surv(time,status!=0)~sex,data=pbc)
publish(sd)
publish(sd,digits=c(3,2))

}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
