% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{subgroupAnalysis}
\alias{subgroupAnalysis}
\title{Subgroup Analysis - Interactions and estimates}
\usage{
subgroupAnalysis(object,dat,timevar=NULL,treatment,
   subgroup, confint.method="default",factor.reference="extraline")
}
\arguments{
\item{object}{- glm, coxph or cph object for which subgroups should be
analyzed.}

\item{dat}{- Dataset including all relevant variables}

\item{timevar}{- Must contain the variable showing observation time for for 
a poisson model of rates - the program used the logarithm of time as
offset, so do not use the logarithm function before the current function.
When not relevant, leave the default value of NULL.}

\item{treatment}{- The variable to be examined in subgroups, best coded
as 0/1 numeric.}

\item{subgroup}{- A set of variables, a string of variables names, presen-
ting the variables where subgroups should be formed. These variables should
all be "factor"}

\item{confint.method}{- "default" creates Wald type confidence interval, "robust",
creates creates robust standard errors - see regressionTable function.}

\item{factor.reference}{- "extraline" creates an extraline for the reference,
"inline" avoids this line.}
}
\value{
A data.frame with subsgroup specifications, number in each subgroup,
parameter estimates and p for interaction.  If a forest plot is further
planned this can be added with "plotConfidence"
}
\description{
In many randomised and observational studies a set of analyses by subgroups
is performed. In the subgroups one variable is of particular interest and it
is requested to tabulate parameter estimates of that variable in subgroups as
well as the probability of interaction.  This function can generate a table
where rows represent sequential subgroups by a range of variables as well as
parameter estimates for a variable of particular interest - and the probabi-
lity of interaction.  The function can examine Cox-models, logistic models 
and Poisson models.
}
\details{
The function can only hand a bivariate outcome, most conviniently coded as
zero or one.  The probability of interaction is from a likelihood ratio test
of nested models with and without the subgroup identifier of interest. Esti-
mates of subgroup effects are from the regressionTable funtion.
}
\examples{
#load libraries
library(data.table)
library(Publish)
library(survival)
data(traceR) #get dataframe traceR
data.table::setDT(traceR)
traceR[,':='(wmi2=factor(wallMotionIndex<0.9,levels=c(TRUE,FALSE), 
                labels=c("bad","good")),
             abd2=factor(abdominalCircumference<95, levels=c(TRUE,FALSE), 
                labels=c("slim","fat")))]
traceR[,sex:=as.factor(sex)] # all subgroup variables needs to be factor                
traceR[observationTime==0,observationTime:=1]           
# univariate analysis of smoking in subgroups of age and sex
# Basic model from randomised study - but observed for 12 years
fit_cox <- coxph(Surv(observationTime,dead)~treatment,data=traceR)
# Selected subgroups - univariable analysis
sub_cox <- subgroupAnalysis(fit_cox,traceR,treatment="treatment",
  subgroup=c("smoking","sex","wmi2","abd2")) # subgroups as character string
# When both start and end are in the Surv statement:
traceR[,null:=0]
fit_cox2 <- coxph(Surv(null,observationTime,dead)~treatment,data=traceR)
summary(regressionTable(fit_cox))
sub_cox2 <- subgroupAnalysis(fit_cox2,traceR,treatment="treatment",
  subgroup=c("smoking","sex","wmi2","abd2")) # subgroups as character string
# Analysis with Poisson - and the unrealistic assumption of constant hazard
# and adjusted for age in all subgroups
fit_p <- glm(dead~treatment+age,family="poisson", 
  offset=log(observationTime),data=traceR)
sub_pois <- subgroupAnalysis(fit_p,traceR,treatment="treatment",
  timevar="observationTime",
  subgroup=~smoking+sex+wmi2+abd2) 
# Analysis with logistic regression - and very wrongly ignoring censoring
fit_log <- glm(dead~treatment+age,family="binomial",data=traceR)
sub_log <- subgroupAnalysis(fit_log,traceR,treatment="treatment",timevar=NULL,
   subgroup=~smoking+sex+wmi2+abd2, factor.reference="inline")
}
\seealso{
coxph, glm, plotConfidence
}
\author{
Christian Torp-Pedersen
}
