% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.regressionTable.R
\name{summary.regressionTable}
\alias{summary.regressionTable}
\title{Formatting regression tables}
\usage{
\method{summary}{regressionTable}(object, showMissing = "ifany",
  print = TRUE, ...)
}
\arguments{
\item{showMissing}{Decide if number of missing values are shown.
Either logical or character. If \code{'ifany'} then number missing values are
shown if there are some.}

\item{print}{If \code{TRUE} print results.}

\item{...}{Used to control formatting of parameter estimates,
confidence intervals and p-values. See examples.}

\item{x}{object obtained with \code{regressionTable}.}
}
\value{
Formatted regression table with raw values as attributes
}
\description{
Preparing regression results for publication
}
\examples{
library(survival)
data(pbc)
pbc$edema <- factor(pbc$edema,levels=c("0","0.5","1"),labels=c("0","0.5","1"))
fit = coxph(Surv(time,status!=0)~age+sex+edema+log(bili)+log(albumin)+log(protime),
            data=pbc)
summary(regressionTable(fit))
summary(regressionTable(fit),handler="prettyNum")
summary(regressionTable(fit),handler="format")
summary(regressionTable(fit),handler="sprintf",digits=c(2,2),pValue.stars=TRUE)
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
publish.glm publish.coxph
}

