% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.subgroupAnalysis.R
\name{summary.subgroupAnalysis}
\alias{summary.subgroupAnalysis}
\title{summary.subgroupAnalysis}
\usage{
table <- function(fit,subgroupVar="subgroupVar",SubgroupVal="SubgroupVal",
Treatment="Treatment", digits=3,eps=0.001,subGroupP=FALSE,keepDigital=FALSE )
}
\arguments{
\item{fit}{- a subgroupAnalysis object}

\item{subgroupVar}{- name for column with variables forming subgroups}

\item{SubgroupVal}{- name of column for levels of subgroupVar}

\item{Treatment}{- name of variable that defines subgroup levels}

\item{digits}{- number of digits for risk ratios and confidence limits}

\item{eps}{- lowest value of p to be shown exactly, others will be "<eps"}
}
\value{
A data.frame with formatted values for subgroups
}
\description{
This function operates on a "subgroupAnalysis" object to produce a formatted
table.
}
\details{
This function produces a formatted table of a subgroupAnalysis object.
A forest plot can be added with the plot function.
}
\examples{
#load libraries
library(Publish)
library(survival)
data(traceR) #get dataframe traceR
traceR$wmi2 <- ifelse(traceR$wallMotionIndex<0.9,"bad","good")
traceR$abd2 <- ifelse(traceR$abdominalCircumference<95,"slim","fat")
# Selected subgroups - univariable analysis
sub <- subgroupAnalysis(traceR,outcome="dead",timevar="observationTime",
  treatment="treatment", subgroup=c("smoking","sex","wmi2","abd2")) 
sub2 <- summary(sub)  
sub2 
}
\seealso{
subgroupAnalysis
}
\author{
Christian Torp-Pedersen
}
