% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sutable.R
\name{sutable}
\alias{sutable}
\title{Fast summary of a univariate table}
\usage{
sutable(...)
}
\arguments{
\item{...}{Unnamed arguments and are passed to \code{univariateTable} as well as named arguments
that match \code{univariateTable}'s arguments, other arguments
are passed to \code{summary.univariateTable}}
}
\value{
Summary table
}
\description{
First apply univariateTable then call summary.
}
\examples{
data(Diabetes)
sutable(gender~age+location+Q(BMI)+height+weight,data=Diabetes,BMI="Body mass index (kg/m^2)")
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
summary.univariateTable univariateTable
}

