% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2x2.R
\name{table2x2}
\alias{table2x2}
\title{2x2 table calculus for teaching}
\usage{
table2x2(
  x,
  digits = 1,
  conf.level = 0.95,
  stats = c("table", "rd", "rr", "or", "chisq", "fisher")
)
}
\arguments{
\item{x}{2x2 table}

\item{digits}{rounding digits}

\item{conf.level}{Confidence level used for constructing confidence intervals. Default is 0.95.}

\item{stats}{subset or all of \code{c("table","rd","or","rr","chisq","fisher")} where rd= risk difference, rr = risk ratio, or = odds ratio, chisq = chi-square test, fisher= fisher's exact test and table = the 2x2 table}
}
\value{
see example
}
\description{
2x2 table calculus for teaching
}
\details{
2x2 table calculus for teaching
}
\examples{
table2x2(table("marker"=rbinom(100,1,0.4),"response"=rbinom(100,1,0.1)))
table2x2(matrix(c(71,18,38,8),ncol=2),stats="table")
table2x2(matrix(c(71,18,38,8),ncol=2),stats=c("rr","fisher"))
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
