% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/publish-package.R
\docType{data}
\name{trace}
\alias{trace}
\title{trace data}
\format{A data frame with 1832 observations on the following 6 variables.
\describe{
\item{Time}{dont know}
\item{smoking}{Smoking status. A factor with levels (Never, Current, Previous)}
\item{sex}{A factor with levels (Female, Male)}
\item{age}{Age in years}
\item{ObsTime}{Observed time at risk}
\item{dead}{Count of deaths}
}}
\source{
CTP
}
\description{
These data are from Christian Torp-Pedersen
}
\examples{
data(trace)
Units(trace,list("age"="years"))
fit <- glm(dead ~ smoking+sex+age+Time+offset(log(ObsTime)), family="poisson",data=trace)
rtf <- regressionTable(fit,factor.reference = "inline")
summary(rtf)
publish(fit)
}
\references{
XXX
}
\keyword{datasets}

