% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish-package.R
\docType{data}
\name{CiTable}
\alias{CiTable}
\title{CiTable data}
\format{
A data frame with 27 observations on the following 9 variables.
\describe{
\item{Drug}{}
\item{Time}{}
\item{Drug.Time}{}
\item{Dose}{}
\item{Mean}{}
\item{SD}{}
\item{n}{}
\item{HazardRatio}{}
\item{lower}{}
\item{upper}{}
\item{p}{}
}
}
\description{
These data are used for testing Publish package functionality.
}
\examples{

data(CiTable)
labellist <- split(CiTable[,c("Dose","Mean","SD","n")],CiTable[,"Drug"])
labellist
plotConfidence(x=CiTable[,c("HazardRatio","lower","upper")], labels=labellist)


}
\keyword{datasets}
