% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/publish.coxph.R
\name{publish.coxph}
\alias{publish.coxph}
\title{Tabulize hazard ratios with confidence intervals and p-values.}
\usage{
\method{publish}{coxph}(object, confint.method, pvalue.method, digits = c(2,
  4), eps = 0.0001, print = TRUE, ci.format = NULL,
  factor.reference = "extraline", units = NULL, ...)
}
\arguments{
\item{object}{A \code{coxph} object.}

\item{digits}{Rounding digits for all numbers but the p-values.}

\item{eps}{passed to format.pval}

\item{print}{If \code{FALSE} do not print results.}

\item{ci.format}{Format for confidence intervals passed to
\code{\link{sprintf}} with two arguments: the lower and the upper
limit.}

\item{...}{passed to labelUnits}

\item{pvalue.digits}{Rounding digits for p-values.}

\item{pvalue.stars}{If \code{TRUE} significance is indicated by
stars instead of p-values.}

\item{showMissing}{If \code{TRUE} show number of missing values in
table}

\item{output.columns}{Select output columns}

\item{reference}{Style for showing results for categorical
variables. If \code{"extraline"} show an additional line for the
reference category.}
}
\value{
Table with hazard ratios, confidence intervals and p-values.
}
\description{
Tabulize the part of the result of a Cox regression analysis which is commonly shown in publications.
}
\details{
Transforms the log hazard ratios to hazard ratios and returns them with confidence limits and p-values.
If explanatory variables are log transformed or log2 transformed, a scaling factor is multiplied to both the log-hazard
ratio and its standard-error.
}
\examples{
library(survival)
data(pbc)
pbc$edema <- factor(pbc$edema,levels=c("0","0.5","1"),labels=c("0","0.5","1"))
fit = coxph(Surv(time,status!=0)~age+sex+edema+log(bili)+log(albumin)+log(protime),
            data=pbc)
publish(fit)

publish(fit,ci.digits=2,pvalue.eps=0.01,pvalue.digits=2,pvalue.stars=TRUE)
publish(fit,ci.digits=2,ci.handler="prettyNum",pvalue.eps=0.01,pvalue.digits=2,pvalue.stars=TRUE)
publish(fit,ci.digits=2,ci.handler="sprintf",pvalue.eps=0.01,pvalue.digits=2,pvalue.stars=TRUE)

fit2 = coxph(Surv(time,status!=0)~age+sex+edema+log(bili,base=2)+log(albumin)+log(protime),
    data=pbc)
publish(fit2)

# with cluster variable
fit3 = coxph(Surv(time,status!=0)~age+cluster(sex)+edema+log(bili,base=2)+log(albumin)+log(protime),
    data=pbc)
publish(fit3)

# with strata and cluster variable
fit4 = coxph(Surv(time,status!=0)~age+cluster(sex)+strata(edema)+log(bili,base=2)+log(albumin)+log(protime),
    data=pbc)
publish(fit4)
}
\author{
Thomas Alexander Gerds
}

