% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.glm.R
\name{publish.glm}
\alias{publish.glm}
\title{Tabulize regression coefficients with confidence intervals and p-values.}
\usage{
\method{publish}{glm}(object, confint.method, pvalue.method, digits = c(2, 4),
  print = TRUE, factor.reference = "extraline", units = NULL, ...)
}
\arguments{
\item{object}{A \code{glm} object.}

\item{confint.method}{See \code{regressionTable}.}

\item{pvalue.method}{See \code{regressionTable}.}

\item{digits}{Rounding digits for all numbers but the p-values.}

\item{print}{If \code{FALSE} do not print results.}

\item{factor.reference}{See \code{regressionTable}.}

\item{units}{See \code{regressionTable}.}

\item{...}{passed to \code{summary.regressionTable} and also
to \code{labelUnits}.}

\item{reference}{Style for showing results for categorical
variables. If \code{"extraline"} show an additional line for the
reference category.}
}
\value{
Table with regression coefficients, confidence intervals and p-values.
}
\description{
Tabulate the results of a generalized linear regression analysis.
}
\details{
The table shows changes in mean for linear regression and
odds ratios for logistic regression (family = binomial).
}
\examples{
data(Diabetes)
## Linear regression
f = glm(bp.2s~frame+gender+age,data=Diabetes)
publish(f)
publish(f,factor.reference="inline")
publish(f,pvalue.stars=TRUE)
publish(f,ci.format="(l,u)")

### interaction
fit = glm(bp.2s~frame+gender*age,data=Diabetes)
summary(fit)
publish(fit)

Fit = glm(bp.2s~frame*gender+age,data=Diabetes)
publish(Fit)

## Logistic regression
Diabetes$hyper1 <- factor(1*(Diabetes$bp.1s>140))
lrfit <- glm(hyper1~frame+gender+age,data=Diabetes,family=binomial)
publish(lrfit)

### interaction
lrfit1 <- glm(hyper1~frame+gender*age,data=Diabetes,family=binomial)
publish(lrfit1)

lrfit2 <- glm(hyper1~frame*gender+age,data=Diabetes,family=binomial)
publish(lrfit2)

## Poisson regression
data(trace)
trace <- Units(trace,list("age"="years"))
fit <- glm(dead ~ smoking+sex+age+Time+offset(log(ObsTime)), family="poisson",data=trace)
rtf <- regressionTable(fit,factor.reference = "inline")
summary(rtf)
publish(fit)

}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}

