% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/publish.glm.R
\name{publish.glm}
\alias{publish.glm}
\title{Tabulize regression coefficients with confidence intervals and p-values.}
\usage{
\method{publish}{glm}(object, confint.method, pvalue.method, digits = c(2, 4),
  eps = 0.0001, print = TRUE, ci.format = NULL,
  factor.reference = "extraline", units = NULL, ...)
}
\arguments{
\item{object}{A \code{glm} object.}

\item{digits}{Rounding digits for all numbers but the p-values.}

\item{eps}{passed to format.pval}

\item{print}{If \code{FALSE} do not print results.}

\item{ci.format}{Format for confidence intervals passed to
\code{\link{sprintf}} with two arguments: the lower and the upper
limit.}

\item{...}{passed to labelUnits}

\item{pvalue.digits}{Rounding digits for p-values.}

\item{pvalue.stars}{If \code{TRUE} significance is indicated by
stars instead of p-values.}

\item{showMissing}{If \code{TRUE} show number of missing values in
table}

\item{output.columns}{Select which parameters of the result are
shown in the table. Defaults to
\code{c("Coefficient","CI.95","pValue","Missing")} for linear
regression and to \code{c("OddsRatio","CI.95","pValue","Missing")}
for logistic regression. Can also include \code{StandardError}.}

\item{intercept}{If \code{FALSE} suppress intercept}

\item{transform}{Transformation for regression coefficients.}

\item{profile}{For logistic regression only. If \code{FALSE} run
with Wald confidence intervals instead of waiting for profile
confidence intervals.}

\item{reference}{Style for showing results for categorical
variables. If \code{"extraline"} show an additional line for the
reference category.}
}
\value{
Table with regression coefficients, confidence intervals and p-values.
}
\description{
Tabulate the results of a generalized linear regression analysis.
}
\details{
The table shows changes in mean for linear regression and
odds ratios for logistic regression (family = binomial).
}
\examples{
data(Diabetes)
f = glm(bp.2s~frame+gender+age,data=Diabetes)
publish(f)
publish(f,reference="inline")
publish(f,pvalue.stars=TRUE)
publish(f,ci.format="(\%1.1f,\%1.1f)")

### interaction
fit = glm(bp.2s~frame+gender*age,data=Diabetes)
summary(fit)
publish(fit)

Fit = glm(bp.2s~frame*gender+age,data=Diabetes)
publish(Fit)

## Logistic regression
Diabetes$hyper1 <- factor(1*(Diabetes$bp.1s>140))
lrfit <- glm(hyper1~frame+gender+age,data=Diabetes,family=binomial)
publish(lrfit)

### interaction
lrfit1 <- glm(hyper1~frame+gender*age,data=Diabetes,family=binomial)
publish(lrfit1)

lrfit2 <- glm(hyper1~frame*gender+age,data=Diabetes,family=binomial)
publish(lrfit2)
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}

