% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{subgroupAnalysis}
\alias{subgroupAnalysis}
\title{Subgroup Analysis - Interactions and estimates}
\usage{
subgroupAnalysis <- function(dat,outcome,timevar=NULL,treatment,
basevars=NULL, subgroup, model="cox",confint="default",factor.reference=
"inline")
}
\arguments{
\item{dat}{- Dataset including all relevant variables}

\item{outcome}{- Variable indicating outcome, mut be numeric zero or one}

\item{timevar}{- Must contain the variable showing time from exposure to
outcome or censoring in case of a cox-model. For a poisson model of rates
this variable is also necessary - the program used the logarithm of time as
offset, so do not use the logarithm function before the curren function.
When not relevant, a value of NULL is appropriate.}

\item{treatment}{- The variable to be examined in subgroups}

\item{basevars}{- A set of variables, a string of variables names, to be 
included in all models. Should NOT include any variables from "subgroup".}

\item{subgroup}{- A set of variables, a string of variables names, presen-
ting the variables where subgroups should be formed.}

\item{model}{- The model to be used, "cox","logistic", "poisson"}

\item{factor.reference}{- "extraline" creates an extraline for the reference,
"inline" avoids this line.}

\item{confint}{- "default" creates Wald type confidence interval, "robust",
creates creates robust standard errors - se function regressionTable.}
}
\value{
A data.frame with subsgroup specifications, number in each subgroup,
parameter estimates and p for interaction.  If a forest plot is further
planned this can be added with "plotConfidence"
}
\description{
In many randomised and observational studies a set of analyses by subgroups
is performed. In the subgroups one variable is of particular interest and it
is requested to tabulate parameter estimates of that variable in subgroups as
well as the probability of interaction.  This function can generate a table
where rows represent sequential subgroups by a range of variables as well as
parameter estimates for a variable of particular interest - and the probabi-
lity of interaction.  The function can examine Cox-models, logistic models 
and Poisson models.
}
\details{
The function selects complete cases based on outcome, treatment and basevars-
and timevar when relevant. For each variable in subgroup complete cases are 
examined. Parameter estimates and p-values for interaction are those calcu-
lated by including interaction terms in "phreg" and "glm".
}
\examples{
#load libraries
library(MASS)
library(data.table)
library(Publish)
library(survival)
data(traceR) # get dataframe traceR
traceR$wmi2 <- ifelse(traceR$wallMotionIndex<0.9,"bad","good")
traceR$abd2 <- ifelse(traceR$abdominalCircumference<95,"slim","fat")
# univariate analysis of smoking in subgroups of age and sex
# Basic model from randomised study - but observed for 12 years
summary(regressionTable(coxph(Surv(observationTime,dead)~treatment,
   data=traceR)))
# Selected subgroups - univariable analysis
subgroupAnalysis(traceR,outcome="dead",timevar="observationTime",
  treatment="treatment", subgroup=c("smoking","sex","wmi2","abd2")) 
# Subgroups but with adjustment for two other variables
subgroupAnalysis(traceR,outcome="dead",timevar="observationTime",
  treatment="treatment", subgroup=c("smoking","sex","wmi2","abd2"),
  basevars=c("age","seCreatinine"))   
# Analysis with Poisson - and the unrealistic assumption of constant hazard
subgroupAnalysis(traceR,outcome="dead",timevar="observationTime",
  treatment="treatment", subgroup=c("smoking","sex","wmi2","abd2"),
  basevars=c("age","seCreatinine"),model="poisson")       
# Analysis with logistic regression - and very wrongly ignoring censoring
subgroupAnalysis(traceR,outcome="dead",timevar=NULL,
  treatment="treatment", subgroup=c("smoking","sex","wmi2","abd2"),
  basevars=c("age","seCreatinine"),model="logistic",factor.reference="inline")   
   
}
\seealso{
coxph, glm, plotConfidence
}
\author{
Christian Torp-Pedersen
}
