% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.regressionTable.R
\name{summary.regressionTable}
\alias{summary.regressionTable}
\title{formatting regression tables}
\usage{
\method{summary}{regressionTable}(x, digits = 2, nsmall = 2,
  handler = "sprintf", print = TRUE, ...)
}
\arguments{
\item{x}{}

\item{digits}{}

\item{nsmall}{}

\item{print}{}

\item{...}{}
}
\value{
Formatted regression table with raw values as attributes
}
\description{
Preparing regression results for publication
}
\examples{
library(survival)
data(pbc)
pbc$edema <- factor(pbc$edema,levels=c("0","0.5","1"),labels=c("0","0.5","1"))
fit = coxph(Surv(time,status!=0)~age+sex+edema+log(bili)+log(albumin)+log(protime),
            data=pbc)
summary(regressionTable(fit))
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
\seealso{
publish.glm publish.coxph
}

